Moved addres + async data fetch logic into App
