+++ src/js/components/App.vue
@@ -88,12 +88,13 @@
 </template>
 
 <script>
-import { ga, formatPrice, initAutoCompleteFields } from '../utils';
+import { ga, formatPrice, initAutoCompleteFields, getNoiseLevelAsText } from '../utils';
 import RR from '../rr';
 import { GMAPS_API_KEY, MAPS_URL } from '../rr';
 import { setPOIs, getPOIs } from '../services/storage';
 import AvailabilityComponent from './Availability.vue';
 import { extractors as pageDataExtractor } from "../sites/index";
+import { loadLocation, loadNoise, loadAirPollution, loadTags, loadParkingZones } from "../services/api";
 
 export default {
   name: 'app',
@@ -145,6 +146,63 @@ export default {
     getPOIs()
       .then(pois => { this.$data.pois = pois; }) // must be in curly braces because of sideEffect
       .catch(err => console.error('An error occurred during POI load from storage' + err.message));
+
+    loadLocation(this.address)
+      .then(({ results }) => {
+        const location = results[0].geometry.location;
+        RR.logDebug('geocoding api response: ', location);
+
+        loadNoise(location, 'day')
+          .then((result) => {
+            RR.logDebug('Noise during the day response: ', result);
+            this.$data.noiseLevel.day = this.$t(getNoiseLevelAsText(result));
+          });
+
+        loadNoise(location, 'night')
+          .then((result) => {
+            RR.logDebug('Noise during the night response: ', result);
+            this.$data.noiseLevel.night = this.$t(getNoiseLevelAsText(result));
+          });
+
+        loadAirPollution(location)
+          .then((airPollutionApiResult) => {
+            RR.logDebug('Air pollution api response: ', airPollutionApiResult);
+            // Definice: Klasifikace klimatologické charakteristiky
+            // 1 = velmi dobrá 2 = dobrá 3 = přijatelná 4 = zhoršená 5 = špatná
+            this.$data.airQuality = this.$t('pollution.value.val' + airPollutionApiResult.value);
+          });
+
+        // TODO vire: map to array of promises and resolve at once
+        // tags
+        loadTags('night_club', location, 500, 2, this);
+        loadTags('transit_station', location, 400, 3, this);
+        loadTags('park', location, 600, 0, this);
+        loadTags('school', location, 1000, 2, this);
+        loadTags('restaurant', location, 500, 3, this);
+
+        loadParkingZones(location, 1000)
+          .then(zones => {
+            if (zones.length > 0) {
+              /*
+               Description of type values  see on the very end of the page
+               http://www.geoportalpraha.cz/cs/fulltext_geoportal?id=BBDE6394B0E14E8BA656DD69CA2EB0F8#.V_Da1HV97eR
+               */
+              const closeBlueZones = zones.filter(pz => {
+                return pz.dist <= 100 /*m*/ && pz.type === 'M';
+                /* Modra  blue zone = parking only for residents */
+              });
+              if (closeBlueZones.length > 0) {
+                this.tags += '<span class="tag" title="' + this.$t('tags.resident_parking.desc') + '">' +
+                  this.$t('tags.resident_parking.title') + '</span>';
+
+                if (zones.filter(pz => pz.dist < 600 && pz.type !== 'M').length > 0) {
+                  this.tags += '<span class="tag" title="' + this.$t('tags.paid_parking.desc') + '">' +
+                    this.$t('tags.paid_parking.title') + '</span>';
+                }
+              }
+            }
+          }); // parking zones
+      });
 
     initAutoCompleteFields(MAPS_URL, GMAPS_API_KEY);
   }
