+++ src/js/contentscript.js
@@ -5,8 +5,7 @@ import VueI18n from 'vue-i18n';
 import RR from './rr';
 import RRLocales from './i18n/locales.js';
 import { extractors as pageDataExtractor } from './sites/index';
-import { streetNamePredicate, addStyles, getNoiseLevelAsText } from './utils';
-import { loadParkingZones, loadNoise, loadAirPollution, loadTags, loadLocation } from './services/api';
+import { streetNamePredicate, addStyles } from './utils';
 
 import '../css/cssnormalize.scss';
 import '../css/panel.scss';
@@ -59,7 +58,8 @@ const loadPanel = function(address) {
         }
       });
 
-      const root = new Vue({
+      // vire: instance is store nowhere
+      new Vue({
         el: '#reality-panel-root',
         render: h => h(App, {
           props: {
@@ -72,64 +72,6 @@ const loadPanel = function(address) {
           }
         })
       });
-      const vueApp = root.$children[0]; // root is a container, and first child is Aapp
-
-      loadLocation(address)
-        .then(({ results }) => {
-          const location = results[0].geometry.location;
-          RR.logDebug('geocoding api response: ', location);
-
-          loadNoise(location, 'day')
-            .then((result) => {
-              RR.logDebug('Noise during the day response: ', result);
-              vueApp.$data.noiseLevel.day = Vue.t(getNoiseLevelAsText(result));
-            });
-
-          loadNoise(location, 'night')
-            .then((result) => {
-              RR.logDebug('Noise during the night response: ', result);
-              vueApp.$data.noiseLevel.night = Vue.t(getNoiseLevelAsText(result));
-            });
-
-          loadAirPollution(location)
-            .then((airPollutionApiResult) => {
-              RR.logDebug('Air pollution api response: ', airPollutionApiResult);
-              // Definice: Klasifikace klimatologické charakteristiky
-              // 1 = velmi dobrá 2 = dobrá 3 = přijatelná 4 = zhoršená 5 = špatná
-              vueApp.$data.airQuality = Vue.t('pollution.value.val' + airPollutionApiResult.value);
-            });
-
-          // TODO vire: map to array of promises and resolve at once
-          // tags
-          loadTags('night_club', location, 500, 2, vueApp);
-          loadTags('transit_station', location, 400, 3, vueApp);
-          loadTags('park', location, 600, 0, vueApp);
-          loadTags('school', location, 1000, 2, vueApp);
-          loadTags('restaurant', location, 500, 3, vueApp);
-
-          loadParkingZones(location, 1000)
-            .then(zones => {
-              if (zones.length > 0) {
-                /*
-                 Description of type values  see on the very end of the page
-                 http://www.geoportalpraha.cz/cs/fulltext_geoportal?id=BBDE6394B0E14E8BA656DD69CA2EB0F8#.V_Da1HV97eR
-                 */
-                const closeBlueZones = zones.filter(pz => {
-                  return pz.dist <= 100 /*m*/ && pz.type === 'M';
-                  /* Modra  blue zone = parking only for residents */
-                });
-                if (closeBlueZones.length > 0) {
-                  vueApp.tags += '<span class="tag" title="' + Vue.t('tags.resident_parking.desc') + '">' +
-                    Vue.t('tags.resident_parking.title') + '</span>';
-
-                  if (zones.filter(pz => pz.dist < 600 && pz.type !== 'M').length > 0) {
-                    vueApp.tags += '<span class="tag" title="' + Vue.t('tags.paid_parking.desc') + '">' +
-                      Vue.t('tags.paid_parking.title') + '</span>';
-                  }
-                }
-              }
-            }); // parking zones
-        }); // geoCode
     });
 }; // loadPanel
