+++ src/js/services/api.js
@@ -34,7 +34,7 @@ export const loadTags = (type, location, radius, minCount, vueApp) => {
       if (response.results.length > minCount) {
         // TODO refactor side effect
         vueApp.tags +=
-          `<span class="tag" title="${Vue.t('tags.' + type + '.desc')}">${Vue.t('tags.' + type + '.title')}</span>`;
+          `<span class="tag" title="${vueApp.$t('tags.' + type + '.desc')}">${vueApp.$t('tags.' + type + '.title')}</span>`;
       }
     });
 };
